+++ android/src/main/kotlin/com/jd/jd_adtiming_plugin/JdAdtimingPlugin.kt
@@ -2,7 +2,6 @@ package com.jd.jd_adtiming_plugin
 
 
 import android.app.Activity
-import android.util.Log
 import androidx.annotation.NonNull
 import com.openmediation.sdk.InitCallback
 import com.openmediation.sdk.InitConfiguration
@@ -42,12 +41,9 @@ class JdAdtimingPlugin : FlutterPlugin, MethodCallHandler, ActivityAware {
 
 
     override fun onMethodCall(@NonNull call: MethodCall, @NonNull result: Result) {
-        Log.e("xxxx123", " method               ${call.method}");
         when (call.method) {
             "init" -> {
                 val appKey = call.argument<String>("appKey") ?: ""
-                Log.e("xxxx123", "appKey :      $appKey")
-
                 val configuration = InitConfiguration.Builder()
                         .appKey(appKey)
                         .logEnable(true)
@@ -55,19 +51,15 @@ class JdAdtimingPlugin : FlutterPlugin, MethodCallHandler, ActivityAware {
 
                 OmAds.init(activity, configuration, object : InitCallback {
                     override fun onSuccess() {
-                        Log.e("xxxx123", " 11>>>>>>        success")
                         channel.invokeMethod("init", "success")
                     }
 
                     override fun onError(result: Error) {
-                        Log.e("xxxx123", " 11>>>>>>        fail")
-
                         channel.invokeMethod("init", "fail")
                     }
                 })
 
-                InterstitialAd.setAdListener(object :
-                        InterstitialAdListener {
+                InterstitialAd.setAdListener(object : InterstitialAdListener {
                     override fun onInterstitialAdAvailabilityChanged(available: Boolean) {
                         if (available) {
                             channel.invokeMethod("interstitialAdListener", "interstitialAdAvailabilityChanged")
@@ -79,7 +71,6 @@ class JdAdtimingPlugin : FlutterPlugin, MethodCallHandler, ActivityAware {
                     }
 
                     override fun onInterstitialAdShowFailed(scene: Scene, error: Error) {
-                        Log.e("xxxx123", "error:          $error")
                         channel.invokeMethod("interstitialAdListener", "interstitialAdShowFailed")
                     }
 
@@ -94,44 +85,36 @@ class JdAdtimingPlugin : FlutterPlugin, MethodCallHandler, ActivityAware {
 
                 RewardedVideoAd.setAdListener(object : RewardedVideoListener {
                     override fun onRewardedVideoAvailabilityChanged(available: Boolean) {
-                        Log.e("xxxx123", "onRewardedVideoAvailabilityChanged :      $available")
                         if (available) {
                             channel.invokeMethod("rewardedVideoAdListener", "rewardedVideoAvailabilityChanged")
                         }
                     }
 
                     override fun onRewardedVideoAdShowed(scene: Scene) {
-                        Log.e("xxxx123", "onRewardedVideoAdShowed")
                         channel.invokeMethod("rewardedVideoAdListener", "rewardedVideoAdShowed")
                     }
 
                     override fun onRewardedVideoAdShowFailed(scene: Scene, error: Error) {
-                        Log.e("xxxx123", "onRewardedVideoAdShowFailed")
                         channel.invokeMethod("rewardedVideoAdListener", "rewardedVideoAdShowFailed")
                     }
 
                     override fun onRewardedVideoAdClicked(scene: Scene) {
-                        Log.e("xxxx123", "onRewardedVideoAdClicked")
                         channel.invokeMethod("rewardedVideoAdListener", "rewardedVideoAdClicked")
                     }
 
                     override fun onRewardedVideoAdClosed(scene: Scene) {
-                        Log.e("xxxx123", "onRewardedVideoAdClosed")
                         channel.invokeMethod("rewardedVideoAdListener", "rewardedVideoAdClosed")
                     }
 
                     override fun onRewardedVideoAdStarted(scene: Scene) {
-                        Log.e("xxxx123", "onRewardedVideoAdStarted")
                         channel.invokeMethod("rewardedVideoAdListener", "rewardedVideoAdStarted")
                     }
 
                     override fun onRewardedVideoAdEnded(scene: Scene) {
-                        Log.e("xxxx123", "onRewardedVideoAdEnded")
                         channel.invokeMethod("rewardedVideoAdListener", "rewardedVideoAdEnded")
                     }
 
                     override fun onRewardedVideoAdRewarded(scene: Scene) {
-                        Log.e("xxxx123", "onRewardedVideoAdRewarded")
                         channel.invokeMethod("rewardedVideoAdListener", "rewardedVideoAdRewarded")
                     }
                 })
@@ -139,12 +122,10 @@ class JdAdtimingPlugin : FlutterPlugin, MethodCallHandler, ActivityAware {
             }
 
             "interstitialAdLoad" -> {
-                Log.e("xxxx123", " 进来               interstitialAdLoad")
                 InterstitialAd.loadAd()
             }
 
             "interstitialShowLoad" -> {
-                Log.e("xxxx123", " interstitialShowLoad               ${InterstitialAd.isReady()}")
                 if (InterstitialAd.isReady()) {
                     InterstitialAd.showAd()
                 }
@@ -160,12 +141,10 @@ class JdAdtimingPlugin : FlutterPlugin, MethodCallHandler, ActivityAware {
 
 
             "rewardedVideoAdLoad" -> {
-                Log.e("xxxx123", " 进来               rewardedVideoAdLoad")
                 RewardedVideoAd.loadAd()
             }
 
             "rewardedVideoShowLoad" -> {
-                Log.e("xxxx123", " rewardedVideoShowLoad               ${RewardedVideoAd.isReady()}")
                 if (RewardedVideoAd.isReady()) {
                     RewardedVideoAd.showAd()
                 }
@@ -173,7 +152,6 @@ class JdAdtimingPlugin : FlutterPlugin, MethodCallHandler, ActivityAware {
 
             "testSuiteLaunch" -> {
                 val appKey = call.argument<String>("appKey") ?: ""
-                Log.e("xxxx123", "testSuiteLaunch :      $appKey")
                 TestSuite.launch(activity, appKey)
             }