+++ lib/jd_adtiming_plugin.dart
@@ -4,10 +4,12 @@ import 'dart:io';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 
+///adtiming插件，这里对接双端api处理
 class JdAdtimingPlugin {
   static const MethodChannel _channel = const MethodChannel('jd_adtiming_plugin');
 
-  //初始化设置
+  ///初始化api
+  ///其中androidAppKey ， iosAppKey 为必须选项且必须适配
   Future<void> init({
     @required String androidAppKey,
     @required String iosAppKey,
@@ -29,7 +31,8 @@ class JdAdtimingPlugin {
   }) async {
     _channel.setMethodCallHandler((MethodCall methodCall) {
       switch (methodCall.method) {
-        //初始化
+
+        ///初始化
         case 'init':
           if (methodCall.arguments == "success") {
             if (success != null) success();
@@ -37,7 +40,8 @@ class JdAdtimingPlugin {
             if (fail != null) fail();
           }
           break;
-        //差评广告监听
+
+        ///差评广告监听
         case 'interstitialAdListener':
           if (methodCall.arguments == "interstitialAdAvailabilityChanged") {
             if (adAvailabilityChanged != null) adAvailabilityChanged();
@@ -51,7 +55,8 @@ class JdAdtimingPlugin {
             if (adClicked != null) adClicked();
           }
           break;
-        //激励广告监听
+
+        ///激励广告监听
         case 'rewardedVideoAdListener':
           if (methodCall.arguments == "rewardedVideoAvailabilityChanged") {
             if (rewardedVideoAvailabilityChanged != null) rewardedVideoAvailabilityChanged();
@@ -80,31 +85,32 @@ class JdAdtimingPlugin {
     _channel.invokeMethod('init', <String, dynamic>{'appKey': Platform.isAndroid ? androidAppKey : iosAppKey});
   }
 
-  //加载插屏广告
+  ///加载插屏广告
   Future<void> interstitialAdLoad() async {
     _channel.invokeMethod('interstitialAdLoad');
   }
 
-  //展示插屏广告
+  ///展示插屏广告
   Future<void> interstitialShowLoad() async {
     _channel.invokeMethod('interstitialShowLoad');
   }
 
-  //加载激励广告
+  ///加载激励广告
   Future<void> rewardedVideoAdLoad() async {
     _channel.invokeMethod('rewardedVideoAdLoad');
   }
 
-  //展示激励广告
+  ///展示激励广告
   Future<void> rewardedVideoShowLoad() async {
     _channel.invokeMethod('rewardedVideoShowLoad');
   }
 
-  //检测插屏广告
+  ///检测插屏广告
   Future<void> checkInterstitialAd({checkInterstitialAdListener: Function}) async {
     _channel.setMethodCallHandler((MethodCall methodCall) {
       switch (methodCall.method) {
-        //初始化
+
+        ///初始化
         case 'checkInterstitialAd':
           if (checkInterstitialAd != null) checkInterstitialAdListener(methodCall.arguments);
           break;
@@ -116,11 +122,12 @@ class JdAdtimingPlugin {
     _channel.invokeMethod('checkInterstitialAd');
   }
 
-  //检测激励广告
+  ///检测激励广告
   Future<void> checkRewardedVideoAd({checkRewardedVideoAdListener: Function}) async {
     _channel.setMethodCallHandler((MethodCall methodCall) {
       switch (methodCall.method) {
-      //初始化
+
+        ///初始化
         case 'checkRewardedVideoAd':
           if (checkInterstitialAd != null) checkRewardedVideoAdListener(methodCall.arguments);
           break;
@@ -132,14 +139,8 @@ class JdAdtimingPlugin {
     _channel.invokeMethod('checkRewardedVideoAd');
   }
 
-
-  //测试结果展示
+  ///测试结果展示
   Future<void> testSuiteLaunch({@required String androidAppKey, @required String iosAppKey}) async {
     _channel.invokeMethod('testSuiteLaunch', <String, dynamic>{'appKey': Platform.isAndroid ? androidAppKey : iosAppKey});
-  }
-
-  static Future<String> get platformVersion async {
-    final String version = await _channel.invokeMethod('getPlatformVersion');
-    return version;
   }
 }